Extract config type boundary
